Seperate arrayParser as a module from main
ArrayParser class를 main으로부터 구분
코드 컨벤션에 맞게 수정, 불필요한 코드 제거
