+++ arrayParser.js
@@ -1,31 +1,23 @@
-const readline = require('readline')
-
-const rl = readline.createInterface({
-  input : process.stdin,
-  output : process.stdout
-});
-
-
-
 class ArrayParser{
-  constructor(){
+  constructor () {
 
   }
   arrayParse(input){
     input = input.trim()
-    if(input[0] === '[' && input[input.length-1] === ']'){  // array
-      let result = {type: 'array', child: []}
-      let token = input.slice(1,input.length-1)
+    let result = {}
+    
+    if (input[0] === '[' && input[input.length-1] === ']') {      // array
+      const tokens = input.slice(1,input.length-1).split(',');
+      result = { type: 'array', child: [] };
       
-      const elements = token.split(',')
-      elements.forEach(element => {
+      tokens.forEach(element => {
         result.child.push(this.arrayParse(element))
       });
       return result
     }
     
-    if(!isNaN((input.trim())*1)){                         //number value 
-      let result = {type: 'number', value: input.trim()}
+    if (!isNaN(input)) {                         //number value 
+      result = { type: 'number', value: input }
       return result
     }
 
@@ -33,16 +25,4 @@ class ArrayParser{
   }
 }
 
-const arrayParser = new ArrayParser();
-
-(() => {
-  rl.question('ArrayParser 사용자 입력 : ', (userInput) => {
-    try{
-      let result = arrayParser.arrayParse(userInput)
-      console.log(JSON.stringify(result, null, 2)); 
-    }catch(e){
-      console.log(e.message)
-    }
-    rl.close();
-  });
-})()
+module.exports = ArrayParser
